Centrality application now has its own SafegraphCBGMapper class.
Cleaned up some ODR violations and refactored some of our helper methods into their own source file.
