+++ src/centrality/centrality.cpp
@@ -4,7 +4,7 @@
 
 
 #include "config.cpp"
-#include "SafegraphMapper.hpp"
+#include "SafegraphCBGMapper.hpp"
 #include "SafegraphTiler.hpp"
 #include <io/parquet.hpp>
 #include <hpx/program_options.hpp>
@@ -34,7 +34,7 @@ using namespace std;
 
 // Register the map-tile instance
 typedef vector<cbg_centrality> reduce_type;
-REGISTER_MAPPER(v2, mt::coordinates::Coordinate3D, SafegraphMapper, SafegraphTiler, reduce_type, string,
+REGISTER_MAPPER(v2, mt::coordinates::Coordinate3D, SafegraphCBGMapper, SafegraphTiler, reduce_type, string,
                 mt::io::FileProvider);
 
 int hpx_main(hpx::program_options::variables_map &vm) {
@@ -91,7 +91,7 @@ int hpx_main(hpx::program_options::variables_map &vm) {
     }
 
     // Initialize all the locales
-    vector<mt::client::MapTileClient<v2, Coordinate3D, SafegraphMapper, SafegraphTiler, reduce_type>> servers;
+    vector<mt::client::MapTileClient<v2, Coordinate3D, SafegraphCBGMapper, SafegraphTiler, reduce_type>> servers;
 
     // Partition the input files, try one for each tile
     const auto csv_path = shared::DirectoryUtils::build_path(config.data_dir, config.patterns_csv);
@@ -125,7 +125,7 @@ int hpx_main(hpx::program_options::variables_map &vm) {
 
                 const auto tile = get<1>(pair);
                 spdlog::debug("Creating server on locale {}", get<0>(pair));
-                mt::client::MapTileClient<v2, Coordinate3D, SafegraphMapper, SafegraphTiler, reduce_type> server(
+                mt::client::MapTileClient<v2, Coordinate3D, SafegraphCBGMapper, SafegraphTiler, reduce_type> server(
                         get<0>(pair), locator,
                         tile,
                         config_values,
