+++ src/components/include/components/BaseSafegraphMapper.hpp
@@ -12,11 +12,24 @@
 #include <map-tile/coordinates/Coordinate3D.hpp>
 #include <shared/ConversionUtils.hpp>
 #include <shared/DateUtils.hpp>
-#include <shared/data.hpp>
 #include <string_view>
 #include <vector>
 
 namespace components {
+
+    weekly_pattern parse_string(std::string_view v);
+
+    size_t compute_temporal_offset(const date::sys_days &start_date, const date::sys_days &row_date);
+
+    std::vector<v2>
+    compute_distance(std::shared_ptr<joined_location> loc, const std::vector<v2> &patterns,
+                     const absl::flat_hash_map<std::string, OGRPoint> &centroids);
+
+    std::vector<v2>
+    expandRow(const weekly_pattern &row,
+              const std::vector<std::pair<std::string, std::uint16_t>> &cbg_visits);
+
+    std::vector<std::pair<std::string, std::uint16_t>> extract_cbg_visits(const weekly_pattern &row);
 
     template<class Backend, class Output>
     class BaseSafegraphMapper {
@@ -56,6 +69,7 @@ namespace components {
                     components::detail::OffsetCalculator(_s->build_offsets().get(), _tc));
 
             // Call nested setup
+            static_cast<Backend &> (*this).setup_impl(ctx);
             spdlog::debug("Mapper setup complete.");
         }
         void map(const mt::ctx::MapContext<Output, mt::coordinates::Coordinate3D> &ctx, const std::string &info) const {
@@ -72,17 +86,44 @@ namespace components {
                 const auto visit_offset = this->_oc->calculate_cbg_offset(r.visit_cbg);
                 if (visit_offset.has_value() && loc_offset.has_value() && g_count >= 0) {
                     // Call nested map
+                    static_cast<Backend const &> (*this).map_impl(ctx, g_count, *loc_offset, *visit_offset, r);
 //                    ctx.emit(mt::coordinates::Coordinate3D(g_count, *loc_offset, *visit_offset), r);
                 }
             });
         }
 
     private:
-        std::vector<v2> process_row(const weekly_pattern &row) const;
-        std::vector<v2> handle_row(const weekly_pattern &row, const std::shared_ptr<joined_location> &jl) const;
-        absl::flat_hash_map<std::string, OGRPoint>
-        get_centroid_map(const std::vector<std::pair<std::string, std::uint16_t>> &visits) const;
-        weekly_pattern parse_string(std::string_view v);
+        [[nodiscard]] std::vector<v2> process_row(const weekly_pattern &row) const {
+            return _l->find_location(row.safegraph_place_id).then(
+                    [this, row](hpx::future<joined_location> location_future) {
+                        const auto jl = std::make_shared<joined_location>(location_future.get());
+                        if (jl->location_cbg.empty()) {
+                            spdlog::warn("Cannot find CBG for safegraph place: {}", row.safegraph_place_id);
+                            const std::vector<v2> f;
+                            return hpx::make_ready_future<std::vector<v2>>(f);
+                        } else {
+                            return hpx::make_ready_future<std::vector<v2>>(handle_row(row, jl));
+                        }
+
+                    }).get();
+        };
+        [[nodiscard]] std::vector<v2> handle_row(const weekly_pattern &row, const std::shared_ptr<joined_location> &jl) const {
+            const auto visits = extract_cbg_visits(row);
+            auto centroid_future = get_centroid_map(visits);
+            std::vector<v2> row_expanded = expandRow(row, visits);
+            return compute_distance(jl, row_expanded, centroid_future);
+        };
+        [[nodiscard]] absl::flat_hash_map<std::string, OGRPoint>
+        get_centroid_map(const std::vector<std::pair<std::string, std::uint16_t>> &visits) const {
+            std::vector<std::string> cbgs;
+            std::transform(visits.begin(), visits.end(), std::back_inserter(cbgs), [](const auto &pair) {
+                return pair.first;
+            });
+
+            const auto centroids = _s->get_centroids(cbgs).get();
+
+            return absl::flat_hash_map<std::string, OGRPoint>(centroids.begin(), centroids.end());
+        }
 
         std::unique_ptr<components::JoinedLocation> _l;
         std::unique_ptr<components::ShapefileWrapper> _s;
